brent-oil-analysis/requirements.txt
# Core data processing
numpy>=1.21.0
pandas>=1.3.0
pandas-datareader>=0.10.0
python-dateutil>=2.8.2

# Bayesian modeling
pymc3>=3.11.5
theano-pymc>=1.1.2
arviz>=0.11.2

# Data visualization
matplotlib>=3.4.0
seaborn>=0.11.0
plotly>=5.3.1

# Web framework
flask>=2.0.0
flask-cors>=3.0.10
gunicorn>=20.1.0

# Data processing and utilities
scikit-learn>=0.24.2
statsmodels>=0.13.0
pytest>=6.2.5

# Jupyter
jupyter>=1.0.0
ipykernel>=6.0.0

# API documentation
flask-restx>=0.5.1

# Environment management
python-dotenv>=0.19.0
